refactor(core): CHECKOUT-9432 Clean up code after experiment removal
