+++ packages/core/auto-export.config.json
@@ -3,16 +3,14 @@
         {
             "inputPath": "packages/**/src/index.ts",
             "ignorePackages": ["test-mocks", "workspace-tools"],
-            "outputPath": "packages/core/src/app/generated/paymentIntegrations/lazy.ts",
-            "memberPattern": ".+PaymentMethod$",
-            "useLazyLoad": true
+            "outputPath": "packages/core/src/app/generated/paymentIntegrations/index.ts",
+            "memberPattern": ".+PaymentMethod$"
         },
         {
             "inputPath": "packages/**/src/index.ts",
             "ignorePackages": ["analytics", "core", "instrument-utils", "locale", "payment-integration-api", "payment-integration-test-framework", "ui", "workspace-tools", "test-utils"],
-            "outputPath": "packages/core/src/app/generated/checkoutButtons/lazy.ts",
-            "memberPattern": ".+Button$",
-            "useLazyLoad": true
+            "outputPath": "packages/core/src/app/generated/checkoutButtons/index.ts",
+            "memberPattern": ".+Button$"
         }
     ],
     "tsConfigPath": "tsconfig.base.json"
