+++ packages/core/src/app/common/resolver/index.ts
@@ -1,2 +1,1 @@
-export { default as resolveComponent } from './resolveComponent';
 export { default as resolveLazyComponent } from './resolveLazyComponent';
