+++ packages/workspace-tools/src/generators/auto-export/auto-export-config.ts
@@ -8,5 +8,4 @@ export interface AutoExportConfigEntry {
     outputPath: string;
     memberPattern: string;
     ignorePackages?: string[];
-    useLazyLoad?: boolean;
 }
