+++ packages/workspace-tools/src/generators/auto-export/auto-export.ts
@@ -11,16 +11,13 @@ export interface AutoExportOptions {
     outputPath: string;
     memberPattern: string;
     tsConfigPath: string;
-    useLazyLoad?: boolean;
 }
 
 export default async function autoExport({
     inputPath,
     ignorePackages,
-    outputPath,
     memberPattern,
     tsConfigPath,
-    useLazyLoad,
 }: AutoExportOptions): Promise<string> {
     let filePaths = await promisify(glob)(inputPath);
 
@@ -32,29 +29,15 @@ export default async function autoExport({
         });
     }
 
-    if (useLazyLoad) {
-        const lazyExports = await createLazyLoadingExports(filePaths, tsConfigPath, memberPattern);
-        const componentRegistry = await createComponentRegistry(filePaths, memberPattern);
+    const lazyExports = await createLazyLoadingExports(filePaths, tsConfigPath, memberPattern);
+    const componentRegistry = await createComponentRegistry(filePaths, memberPattern);
 
-        validateComponentNames(lazyExports, componentRegistry);
+    validateComponentNames(lazyExports, componentRegistry);
 
-        return `${generateLazyLoadingCode(lazyExports)}
+    return `${generateLazyLoadingCode(lazyExports)}
 
 ${generateComponentRegistryCode(componentRegistry)}
-        `;
-    }
-
-    const exportDeclarations = await Promise.all(
-        filePaths.map((filePath) => createExportDeclaration(filePath, tsConfigPath, memberPattern)),
-    );
-
-    return ts
-        .createPrinter()
-        .printList(
-            ts.ListFormat.MultiLine,
-            ts.factory.createNodeArray(exportDeclarations.filter(exists)),
-            ts.createSourceFile(outputPath, '', ts.ScriptTarget.ESNext),
-        );
+    `;
 }
 
 function validateComponentNames(
@@ -71,55 +54,6 @@ function validateComponentNames(
     if (onlyInRegistry.length > 0) {
         throw new Error(`Component name mismatch detected: ${onlyInRegistry.join(', ')}\n`);
     }
-}
-
-async function createExportDeclaration(
-    filePath: string,
-    tsConfigPath: string,
-    memberPattern: string,
-): Promise<ts.ExportDeclaration | undefined> {
-    const root = await getSource(filePath);
-
-    const memberNames = root.statements
-        .filter(ts.isExportDeclaration)
-        .flatMap((statement) => {
-            if (
-                !statement.exportClause ||
-                !ts.isNamedExports(statement.exportClause) ||
-                // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-                !statement.exportClause.elements
-            ) {
-                return [];
-            }
-
-            return statement.exportClause.elements.filter(ts.isExportSpecifier);
-        })
-
-        .map((element) =>
-            'escapedText' in element.name
-                ? element.name.escapedText.toString()
-                : JSON.stringify(element.name),
-        )
-        .filter((memberName: string) => new RegExp(memberPattern).exec(memberName));
-
-    if (memberNames.length === 0) {
-        return;
-    }
-
-    return ts.factory.createExportDeclaration(
-        undefined,
-        false,
-        ts.factory.createNamedExports(
-            memberNames.map((memberName) =>
-                ts.factory.createExportSpecifier(
-                    false,
-                    undefined,
-                    ts.factory.createIdentifier(memberName),
-                ),
-            ),
-        ),
-        ts.factory.createStringLiteral(getImportPath(filePath, tsConfigPath), true),
-    );
 }
 
 async function getSource(filePath: string): Promise<ts.SourceFile> {
@@ -141,10 +75,6 @@ function getImportPath(packagePath: string, tsConfigPath: string): string {
     }
 
     throw new Error('Unable to resolve to a valid package.');
-}
-
-function exists<TValue>(value?: TValue): value is NonNullable<TValue> {
-    return value !== null && value !== undefined;
 }
 
 async function createLazyLoadingExports(
