+++ packages/workspace-tools/src/generators/auto-export/is-auto-export-config.ts
@@ -34,10 +34,6 @@ export default function isAutoExportConfig(config: unknown): config is AutoExpor
             return false;
         }
 
-        if (hasKey(entry, 'useLazyLoad') && typeof entry.useLazyLoad !== 'boolean') {
-            return false;
-        }
-
         return true;
     });
 }
